Deposit is basically done except DT & vision stuff
Everything in Deposit that needs to be done is just setting parameters of stuff & it relies on a method yet to be made in Intake. There's also DT coordinate & vision stuff that I need to figure out.

The other classes is just a bunch of parameter stuff.
I do need to figure out/confirm the servo types for everything and also the backwards orientation settings for the servos and motors.